Commit 7
In this commit the new spot functionality is complete. The user can add
a new location and it will appear on their spots list on the home
screen. The new spot UI needs to be cleaned up but is finished
functionality wise. A database helper class and database adapter class
were created. The helper class to create a local database and deal with
CRUD functions. The adapter to adapt the data to be displayed in the
list view.