Commit 8
In this commit I updated the MainMapFragment so that it will populate
the map with all of spots retrieved from the database. If the user has
just logged in then it adds all the markers to the map then zooms in on
the top spot on their list. If they just added a new spot then all the
markers are added to the map and then the map zooms in on the newly
added spot.