Commit 9
This commit adds additional functionality to the home screen and new
spot screen. Now when the user has selected a location and is giving it
a name and description they can choose to go back and change the
location. The name and description will still be there when they get
back to the screen. On the home screen the user can now click on a list
view item and the map will zoom to that location. When the user is
adding a new location and interacting with the map view, the info
window shows the coordinates of that marker.