Commit 10
In this commit I have added the spot detail fragment, a page that
displays all of the details about that spot. When the user clicks on
the info window of a marker it passes the marker id to the new fragment
and queries the data for that spot. I have also created a new activity
that will be used for adding and viewing catches.