Commit 15
In this commit I created a catch detail fragment and added
functionality that allows the user to click on a catch item in the list
view on the spot details page. This brings them to the catch detail
page. I also set up fire store and when the user adds a new spot it
adds that spot to firebase.